+++ src/com/dci/intellij/dbn/object/DBProgram.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface DBProgram<P extends DBProcedure, F extends DBFunction> extends DBSchemaObject {
     List<P> getProcedures();
     List<F> getFunctions();
-    F getFunction(String name, int overload);
-    P getProcedure(String name, int overload);
-    DBMethod getMethod(String name, int overload);
+    F getFunction(String name, short overload);
+    P getProcedure(String name, short overload);
+    DBMethod getMethod(String name, short overload);
     boolean isEmbedded();
 }
