+++ src/com/dci/intellij/dbn/object/impl/DBSchemaImpl.java
@@ -187,7 +187,7 @@ public class DBSchemaImpl extends DBObjectImpl<DBSchemaMetadata> implements DBSc
     }
 
     @Override
-    public DBObject getChildObject(DBObjectType objectType, String name, int overload, boolean lookupHidden) {
+    public DBObject getChildObject(DBObjectType objectType, String name, short overload, boolean lookupHidden) {
         if (objectType.isSchemaObject()) {
             DBObject object = super.getChildObject(objectType, name, overload, lookupHidden);
             if (object == null && objectType != SYNONYM) {
@@ -392,14 +392,14 @@ public class DBSchemaImpl extends DBObjectImpl<DBSchemaMetadata> implements DBSc
     }
 
     @Override
-    public DBProcedure getProcedure(String name, int overload) {
+    public DBProcedure getProcedure(String name, short overload) {
         return overload > 0 ?
                 procedures.getObject(name, overload) :
                 getObjectFallbackOnSynonym(procedures, name);
     }
 
     @Override
-    public DBFunction getFunction(String name, int overload) {
+    public DBFunction getFunction(String name, short overload) {
         return overload > 0 ?
                 functions.getObject(name, overload) :
                 getObjectFallbackOnSynonym(functions, name);
@@ -413,7 +413,7 @@ public class DBSchemaImpl extends DBObjectImpl<DBSchemaMetadata> implements DBSc
     }
 
     @Override
-    public DBMethod getMethod(String name, DBObjectType methodType, int overload) {
+    public DBMethod getMethod(String name, DBObjectType methodType, short overload) {
         if (methodType == null) {
             DBMethod method = getProcedure(name, overload);
             if (method == null) method = getFunction(name, overload);
@@ -427,7 +427,7 @@ public class DBSchemaImpl extends DBObjectImpl<DBSchemaMetadata> implements DBSc
     }
 
     @Override
-    public DBMethod getMethod(String name, int overload) {
+    public DBMethod getMethod(String name, short overload) {
         return getMethod(name, null, overload);
     }
 
@@ -1039,7 +1039,7 @@ public class DBSchemaImpl extends DBObjectImpl<DBSchemaMetadata> implements DBSc
                 String programName = metadata.getProgramName();
                 String methodName = metadata.getMethodName();
                 String methodType = metadata.getMethodType();
-                int overload = metadata.getOverload();
+                short overload = metadata.getOverload();
                 DBSchema schema = (DBSchema) content.getParentElement();
                 DBProgram program = programName == null ? null : schema.getProgram(programName);
 
@@ -1072,7 +1072,7 @@ public class DBSchemaImpl extends DBObjectImpl<DBSchemaMetadata> implements DBSc
                 String datasetName = metadata.getDatasetName();
                 String columnName = metadata.getColumnName();
                 String constraintName = metadata.getConstraintName();
-                int position = metadata.getPosition();
+                short position = metadata.getPosition();
 
 /*
             DBSchema schema = (DBSchema) dynamicContent.getParent();
